requirements.txt
altair==5.2.0
anyio==4.3.0
attrs==23.2.0
bcrypt==3.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
extra-streamlit-components==0.1.71
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
httpx-oauth==0.13.2
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
patsy==0.5.6
pillow==10.2.0
plotly==5.20.0
plotly-express==0.4.1
protobuf==4.25.3
pyarrow==15.0.2
pycparser==2.21
pydeck==0.8.1b0
PyGithub==2.3.0
pygments==2.17.2
PyJWT==2.8.0
pyngrok==7.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
statsmodels==0.14.1
streamlit==1.32.2
streamlit-authenticator==0.3.1
streamlit-oauth==0.1.8
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0